Update abstract.org
Find my additions here.  We are 1270 word count with this (not including our names and the chart info), so it could use some editing at some point.  But I think it's in good shape to share with Michael Zimmer!  Once I hit "propose changes," if I get to something that looks shareable, I will share the link with Michael Zimmer now.  Otherwise, I will wait for your attention today to get it into a format that's ready to share!

Isabelle